refactor saved state
- create a single saved_state module instead of having two
- export an object with methods instead of a function with methods
