+++ packages/server/lib/gui/windows.coffee
@@ -231,7 +231,7 @@ module.exports = {
       newState[keys.height] = height
       newState[keys.x] = x
       newState[keys.y] = y
-      savedState(projectRoot, isTextTerminal)
+      savedState.create(projectRoot, isTextTerminal)
       .then (state) ->
         state.set(newState)
     , 500
@@ -243,7 +243,7 @@ module.exports = {
       newState = {}
       newState[keys.x] = x
       newState[keys.y] = y
-      savedState(projectRoot, isTextTerminal)
+      savedState.create(projectRoot, isTextTerminal)
       .then (state) ->
         state.set(newState)
     , 500
@@ -251,14 +251,14 @@ module.exports = {
     win.webContents.on "devtools-opened", ->
       newState = {}
       newState[keys.devTools] = true
-      savedState(projectRoot, isTextTerminal)
+      savedState.create(projectRoot, isTextTerminal)
       .then (state) ->
         state.set(newState)
 
     win.webContents.on "devtools-closed", ->
       newState = {}
       newState[keys.devTools] = false
-      savedState(projectRoot, isTextTerminal)
+      savedState.create(projectRoot, isTextTerminal)
       .then (state) ->
         state.set(newState)
