+++ packages/server/test/unit/plugins/preprocessor_spec.coffee
@@ -4,7 +4,7 @@ EE = require("events")
 Fixtures = require("../../support/helpers/fixtures")
 path = require("path")
 appData = require("#{root}../lib/util/app_data")
-{ toHashName } = require("#{root}../lib/util/saved_state")
+savedState = require("#{root}../lib/saved_state")
 
 plugins = require("#{root}../lib/plugins")
 preprocessor = require("#{root}../lib/plugins/preprocessor")
@@ -39,12 +39,12 @@ describe "lib/plugins/preprocessor", ->
 
     it "executes the plugin with output path", ->
       preprocessor.getFile(@filePath, @config)
-      expectedPath = appData.projectsPath(toHashName(@todosPath), "bundles", @filePath)
+      expectedPath = appData.projectsPath(savedState.toHashName(@todosPath), "bundles", @filePath)
       expect(@plugin.lastCall.args[0].outputPath).to.equal(expectedPath)
 
     it "executes the plugin with output path when integrationFolder was defined", ->
       preprocessor.getFile(@integrationFolder + @filePath, Object.assign({integrationFolder: @integrationFolder}, @config))
-      expectedPath = appData.projectsPath(toHashName(@todosPath), "bundles", @filePath)
+      expectedPath = appData.projectsPath(savedState.toHashName(@todosPath), "bundles", @filePath)
       expect(@plugin.lastCall.args[0].outputPath).to.equal(expectedPath)
 
     it "returns a promise resolved with the plugin's outputPath", ->
