+++ packages/server/test/unit/socket_spec.coffee
@@ -13,7 +13,6 @@ Socket       = require("#{root}lib/socket")
 Server       = require("#{root}lib/server")
 Automation   = require("#{root}lib/automation")
 exec         = require("#{root}lib/exec")
-savedState   = require("#{root}lib/saved_state")
 preprocessor = require("#{root}lib/plugins/preprocessor")
 fs           = require("#{root}lib/util/fs")
 open         = require("#{root}lib/util/open")
